fixed race in Pack() on default options initialization. (#67)
* fixed race in Pack() on default options initialization.

Default Options used in Pack() is a global object with zero-initialized
fields. Validate() changes PtrSize field value on the first call
from zero to 32. The data race detector complains if multiple goroutines call Pack().

This commit fixes this issue and adds a tests its correctness. Important
note about test is that it should be in a separate package because
it needs to be run exclusively so that it is guaranteed that it is run
first to catch this race condition.